refactor: Update backend module structure and imports
