+++ src/backend/impl_profile/database.rs
@@ -1,12 +1,13 @@
-use serde::{Deserialize, Serialize};
+use std::cell::RefCell;
 
-use super::Profile;
+use super::{Profile, ProfileType};
+
 pub type ProfileDataBase = std::collections::HashMap<String, ProfileType>;
 
 /// manage profiles
 pub struct ProfileManager {
-    current: core::cell::RefCell<String>,
-    all: core::cell::RefCell<ProfileDataBase>,
+    current: RefCell<String>,
+    all: RefCell<ProfileDataBase>,
 }
 impl ProfileManager {
     pub fn new(current: String, all: ProfileDataBase) -> Self {
@@ -59,30 +60,5 @@ impl ProfileManager {
             "Selected profile not in database"
         );
         *self.current.borrow_mut() = pf.name;
-    }
-}
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub enum ProfileType {
-    /// local import
-    File,
-    /// download url
-    Url(String),
-    /// generated by template
-    Generated(String),
-}
-impl ProfileType {
-    pub fn is_file(&self) -> bool {
-        matches!(self, ProfileType::File)
-    }
-    pub fn is_upgradable(&self) -> bool {
-        !self.is_file()
-    }
-    /// if [ProfileType::is_upgradable], return [Some]
-    pub fn get_domain(&self) -> Option<String> {
-        match self {
-            ProfileType::File => None,
-            ProfileType::Url(url) => crate::clash::util::extract_domain(url).map(|s| s.to_owned()),
-            ProfileType::Generated(name) => Some(format!("From template {name}")),
-        }
     }
 }
