+++ src/backend/impl_profile/profile.rs
@@ -1,9 +1,5 @@
 use std::{fs::File, path::PathBuf};
 
-pub mod map;
-
-pub use map::ProfileType;
-
 #[derive(Clone)]
 pub struct Profile {
     pub name: String,
@@ -114,5 +110,31 @@ impl LocalProfile {
             self.content = serde_yml::from_reader(fp)?;
         }
         Ok(())
+    }
+}
+
+#[derive(Debug, serde::Serialize, serde::Deserialize, Clone)]
+pub enum ProfileType {
+    /// local import
+    File,
+    /// download url
+    Url(String),
+    /// generated by template
+    Generated(String),
+}
+impl ProfileType {
+    pub fn is_file(&self) -> bool {
+        matches!(self, ProfileType::File)
+    }
+    pub fn is_upgradable(&self) -> bool {
+        !self.is_file()
+    }
+    /// if [ProfileType::is_upgradable], return [Some]
+    pub fn get_domain(&self) -> Option<String> {
+        match self {
+            ProfileType::File => None,
+            ProfileType::Url(url) => crate::clash::util::extract_domain(url).map(|s| s.to_owned()),
+            ProfileType::Generated(name) => Some(format!("From template {name}")),
+        }
     }
 }
